# brand-platform/docs/openapi-mobile-v1.yaml
openapi: 3.0.3
info:
  title: CADA Brand Partnerships — Mobile API
  version: 1.0.0
  description: |
    APIs consumed by the CADA iOS app for brand challenge discovery,
    enrollment, habit attribution, and (Phase 5) QR rewards.
servers:
  - url: https://api.cada.app/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: CADA app user JWT (not brand staff)
  schemas:
...
    Enrollment:
      type: object
      properties:
        enrollment_id: { type: string, format: uuid }
        challenge_id: { type: string, format: uuid }
        status:
          type: string
          enum: [active, completed, dropped]
        enrolled_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        completion_count: { type: integer }
        already_enrolled: { type: boolean }
...
paths:
...
  /challenges/{id}/enroll:
    post:
      summary: Enroll in a challenge
      operationId: enrollChallenge
      tags: [Challenges]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201':
          description: Enrolled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '200':
          description: Already enrolled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '410':
          description: Challenge unavailable
...